chore: separate State file
