+++ pkg/protocol/messages.go
@@ -1,9 +1,5 @@
 package protocol
 
-import (
-	"github.com/six78/2-story-points-cli/internal/config"
-)
-
 const Version byte = 1
 
 type PlayerID string
@@ -15,42 +11,6 @@ type Issue struct {
 	Votes      IssueVotes `json:"votes"`
 	Result     *VoteValue `json:"result"` // NOTE: keep pointer. Because "empty string means vote is not revealed"
 	Hint       *Hint      `json:"-"`
-}
-
-type VoteState string
-
-const (
-	IdleState     VoteState = "idle"     // ActiveIssue == ""
-	VotingState   VoteState = "voting"   // ActiveIssue != "", Revealed == false
-	RevealedState VoteState = "revealed" // ActiveIssue != "", Revealed == true, Issues[ActiveIssue].Result == nil
-	FinishedState VoteState = "finished" // ActiveIssue != "", Revealed == true, Issues[ActiveIssue].Result != nil
-)
-
-type State struct {
-	Players       PlayersList `json:"players"`
-	Issues        IssuesList  `json:"issues"`
-	ActiveIssue   IssueID     `json:"activeIssue"`
-	VotesRevealed bool        `json:"votesRevealed"`
-	Timestamp     int64       `json:"-"` // TODO: Fix conflict with Message.Timestamp. Change type to time.Time.
-	Deck          Deck        `json:"-"`
-}
-
-func (s *State) VoteState() VoteState {
-	if s.ActiveIssue == "" {
-		return IdleState
-	}
-	if !s.VotesRevealed {
-		return VotingState
-	}
-	issue := s.Issues.Get(s.ActiveIssue)
-	if issue == nil {
-		config.Logger.Error("active issue not found when calculating vote state")
-		return IdleState
-	}
-	if issue.Result == nil {
-		return RevealedState
-	}
-	return FinishedState
 }
 
 type MessageType string
